+++ src/components/PageNavLink.tsx
@@ -1,0 +1,24 @@
+import classNames from 'classnames';
+import { FC } from 'react';
+import { NavLink } from 'react-router-dom';
+
+type Props = {
+  to: string,
+  text: string,
+};
+
+export const PageNavLink: FC<Props> = ({ to, text }) => {
+  return (
+    <NavLink
+      to={to}
+      className={({ isActive }) => classNames(
+        'navbar-item',
+        {
+          'has-background-grey-lighter': isActive,
+        },
+      )}
+    >
+      {text}
+    </NavLink>
+  );
+};
