+++ src/components/PersonLink.tsx
@@ -1,12 +1,12 @@
-import classNames from "classnames";
-import { FC } from "react";
-import { Link } from "react-router-dom";
+import classNames from 'classnames';
+import { FC } from 'react';
+import { Link } from 'react-router-dom';
 
 type Props = {
   slug: string,
   sex: string,
   name: string,
-}
+};
 
 export const PersonLink: FC<Props> = ({ slug, sex, name }) => (
   <Link
